refactored a bit, se server.py is not to be imported and  GHA doesn't fail.
