+++ src/tests/conftest.py
@@ -3,7 +3,17 @@ import json
 import os
 import pytest
 from lxml import etree
-from server import config_absolute_paths, get_active_profile
+from config_util import config_absolute_paths
+
+
+def get_active_profile(app):
+    def_profile = app.config['PROFILES'][0]
+    for p in app.config['PROFILES']:
+        if 'default' in p and p['default'] is True:
+            def_profile = p
+            break
+    return def_profile
+
 
 """
 Basic fixtures that are useful for most of the test modules
@@ -80,8 +90,8 @@ def application_settings():
     """ Returns the application settings."""
     app = Flask(__name__)
     app.config.from_object('config.settings_example.Config')
-    app.config['ACTIVE_PROFILE'] = get_active_profile()
-    config_absolute_paths()
+    app.config['ACTIVE_PROFILE'] = get_active_profile(app)
+    config_absolute_paths(app)
 
     return app.config
